* added table handling in setup script  * added handling of static data (constants) and restart files to archive script  * revised run_cycle scripts and added archiving of static data  * added scenarion name to CAMtr-files

+++ Scripts/GPC/run_cycle_pbs.sh
@@ -5,37 +5,48 @@
 
 # settings
 set -e # abort if anything goes wrong
-export STEPFILE='stepfile' # file in $INIDIR
-export INIDIR="${PWD}" # current directory
+NOGEO=$1 # option to run without geogrid
+STEPFILE='stepfile' # file in $INIDIR
+INIDIR="${PWD}" # current directory
+METDATA="${INIDIR}/metgrid/"
+WRFOUT="${INIDIR}/wrfout/"
+CASENAME='cycling' # name tag
+WPSSCRIPT="run_${CASENAME}_WPS.pbs"
+WRFSCRIPT="run_${CASENAME}_WRF.pbs"
+STATICTGZ='static.tgz' # file for static data backup
 
 # launch feedback
 echo
 echo "   ***   Starting Cycle  ***   "
 echo
-echo "   Stepfile: ${STEPFILE}"
+# echo "   Stepfile: ${STEPFILE}"
 echo "   Root Dir: ${INIDIR}"
-echo 
+echo
 
 # clear some folders
 cd "${INIDIR}"
-export METDATA="${INIDIR}/metgrid/"
-export WRFOUT="${INIDIR}/wrfout/"
 echo "   Clearing Output Folders:"
 echo "${METDATA}"
 echo "${WRFOUT}"
 rm -rf "${METDATA}" "${WRFOUT}" 
+mkdir -p "${WRFOUT}"
 
 # run geogrid
 # clear files
 cd "${INIDIR}"
-rm -f geo_em.d??.nc geogrid.log*
-# run with parallel processes
-echo
-echo "   Running geogrid.exe"
-echo
-mpirun -n 4 ./geogrid.exe
+if [[ "${NOGEO}" == 'NOGEO'* ]]; then
+  echo "   Not running geogrid.exe"
+else
+  rm -f geo_em.d??.nc geogrid.log*
+  # run with parallel processes
+  echo
+  echo "   Running geogrid.exe"
+  echo
+  mpirun -n 4 ./geogrid.exe
+fi
 
 # read first entry in stepfile 
+export STEPFILE
 NEXTSTEP=$(python cycling.py)
 #export NEXTSTEP
 echo
@@ -48,10 +59,23 @@ sed -i '/restart\s/ s/restart\s*=\s*\.true\..*$/restart = .false.,/' "${INIDIR}/
 # and make sure the rest is on restart
 sed -i '/restart\s/ s/restart\s*=\s*\.false\..*$/restart = .true.,/' "${INIDIR}/namelist.input"
 echo "   Setting restart option and interval in namelist."
+# echo
+
+# create backup of static files
+cd "${INIDIR}"
+rm -rf 'static-tmp/'
+mkdir -p 'static-tmp'
+echo $( cp -P * 'static-tmp/' &> /dev/null ) # trap this error and hide output
+cp -rL 'meta/' 'tables/' 'static-tmp/'
+tar czf "${STATICTGZ}" 'static-tmp/'
+rm -r 'static-tmp/'
+mv "${STATICTGZ}" "${WRFOUT}"
+echo "  Saved backup file for static data:"
+echo "${WRFOUT}/${STATICTGZ}"
 echo
 
-# submit first WPS instance
-qsub ./run_cycling_WPS.pbs -v NEXTSTEP="${NEXTSTEP}"  
-
-# submit first WRF instance
-qsub ./run_cycling_WRF.pbs -v NEXTSTEP="${NEXTSTEP}"
+# # submit first WPS instance
+# qsub ./${WPSSCRIPT} -v NEXTSTEP="${NEXTSTEP}"
+# 
+# # submit first WRF instance
+# qsub ./${WRFSCRIPT} -v NEXTSTEP="${NEXTSTEP}"
